server/requirements.txt
mysql-connector-python
fastapi
uvicorn[standard]
PyJWT
python-multipart